extract cached op
