+++ src/c_api/c_api_ndarray.cc
@@ -36,6 +36,7 @@
 #include "../common/utils.h"
 #include "../common/exec_utils.h"
 #include "../imperative/imperative_utils.h"
+#include "../imperative/cached_op.h"
 
 using namespace mxnet;
 
@@ -160,12 +161,11 @@ int MXCreateCachedOp(SymbolHandle handle,
   std::vector<std::string> input_names;
   input_names.reserve(inputs.size());
   for (const auto& i : inputs) input_names.push_back(i->attrs.name);
-  *out = new std::shared_ptr<Imperative::CachedOp>(
-      new Imperative::CachedOp(
-        *sym,
-        std::vector<std::pair<std::string, std::string> >(),
-        input_names,
-        std::unordered_map<std::string, std::vector<NDArray> >()));
+  *out = new CachedOpPtr(new CachedOp(
+      *sym,
+      std::vector<std::pair<std::string, std::string> >(),
+      input_names,
+      std::unordered_map<std::string, std::vector<NDArray> >()));
   API_END();
 }
 
@@ -195,8 +195,7 @@ int MXCreateCachedOpEx(SymbolHandle handle,
     param_dict[param_names[i]].emplace_back(
         *reinterpret_cast<NDArray*>(params[i]));
   }
-  *out = new std::shared_ptr<Imperative::CachedOp>(
-      new Imperative::CachedOp(*sym, flags, args, param_dict));
+  *out = new CachedOpPtr(new CachedOp(*sym, flags, args, param_dict));
   API_END();
 }
